+++ CHANGELOG.md
@@ -1,7 +1,10 @@
 ## 2.20.1 (Unreleased)
+FEATURES:
+* **New Resource:** sumologic_cse_first_seen_rule (GH-472)
+
 BUG FIXES:
 * Added validations to fail if no trigger conditions defined in monitor resource (GH-463)
-
+ 
 ## 2.20.0 (December 15, 2022)
 FEATURES:
 * **New Resource:** sumologic_cse_custom_match_list_column (GH-462)